+++ test/end-to-end/fixtures/plugin/options.js
@@ -1,0 +1,7 @@
+module.exports = {
+  exampleOption: {
+    type: 'boolean',
+    describe: 'Plugin option example',
+    default: false
+  }
+}