+++ CHANGELOG.md
@@ -6,9 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+- Add plugin Constructor
+- Add events api for plugins
+- Add controller api interface for plugins
+- Expose extendOpenApi and addOperations methods to module and plugin
 ### Changed
 ### Fixed
 ### Removed
+### BREAKING CHANGES
+- Changed Controller uris to adapt them to Controller version 1.0.0-alpha.9
+- Expose event emitter in events object, instead of module object directly
 
 ## [1.0.0-alpha.1] - 2018-11-4
 ### Added